Resolve issue where redirect with a fragment is invalid after appending token
Description:
location is tokenized on '#'. If a fragment exists, it is re-added back to location. Tokenization returns at most 2 substrings. Only the first '#' found in location is the index for tokenization, e.g. /myPath#myFragment#blah becomes [0] /myPath [1] myFragment#blah. Any '#' in the URL prior to fragment should have been encoded as '%23'.